re-organised newsletter modal components
