+++ app/components/modals/newsletters/confirm-create.hbs
