+++ app/components/modals/newsletters/confirm-newsletter-email.hbs
