+++ app/components/modals/newsletters/edit/design.hbs
