+++ app/components/modals/newsletters/edit/design.js
