+++ app/components/modals/newsletters/edit/settings.hbs
