+++ app/components/modals/newsletters/edit/settings.js
