+++ app/routes/settings/members-email-labs.js
@@ -1,6 +1,6 @@
 import AdminRoute from 'ghost-admin/routes/admin';
 import ConfirmUnsavedChangesModal from '../../components/modals/confirm-unsaved-changes';
-import VerifyNewsletterEmail from '../../components/modals/edit-newsletter/verify-newsletter-email';
+import VerifyNewsletterEmail from '../../components/modals/newsletters/verify-newsletter-email';
 import {action} from '@ember/object';
 import {inject as service} from '@ember/service';
